feat: add optimism module and refactor types
