+++ crates/rpc-types/src/eth/transaction/receipt.rs
@@ -55,30 +55,3 @@ pub struct TransactionReceipt {
     #[serde(flatten)]
     pub other: OtherFields,
 }
-
-/// Additional fields for Optimism transaction receipts
-#[derive(Clone, Copy, Default, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct OptimismTransactionReceiptFields {
-    /// Deposit nonce for deposit transactions post-regolith
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deposit_nonce: Option<U64>,
-    /// L1 fee for the transaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub l1_fee: Option<U256>,
-    /// L1 fee scalar for the transaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub l1_fee_scalar: Option<U256>,
-    /// L1 gas price for the transaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub l1_gas_price: Option<U256>,
-    /// L1 gas used for the transaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub l1_gas_used: Option<U256>,
-}
-
-impl From<OptimismTransactionReceiptFields> for OtherFields {
-    fn from(value: OptimismTransactionReceiptFields) -> Self {
-        serde_json::to_value(value).unwrap().try_into().unwrap()
-    }
-}
